TP#19925: Change dhl account number from int to long

+++ Code/ShipWorks.Data.Model/EntityInterfaces/IDhlExpressAccountEntity.cs
@@ -36,9 +36,9 @@ namespace ShipWorks.Data.Model.EntityInterfaces
         /// <summary> The AccountNumber property of the Entity DhlExpressAccount<br/><br/>
         /// </summary>
         /// <remarks>Mapped on table field: "DhlExpressAccount"."AccountNumber"<br/>
-        /// Table field type characteristics (type, precision, scale, length): Int, 10, 0, 0<br/>
+        /// Table field type characteristics (type, precision, scale, length): BigInt, 19, 0, 0<br/>
         /// Table field behavior characteristics (is nullable, is PK, is identity): false, false, false</remarks>
-        System.Int32 AccountNumber { get; }
+        System.Int64 AccountNumber { get; }
         /// <summary> The ShipEngineCarrierId property of the Entity DhlExpressAccount<br/><br/>
         /// </summary>
         /// <remarks>Mapped on table field: "DhlExpressAccount"."ShipEngineCarrierId"<br/>
