+++ Code/ShipWorks.Data.Model/ReadOnlyEntities/ReadOnlyDhlExpressAccountEntity.cs
@@ -75,9 +75,9 @@ namespace ShipWorks.Data.Model.ReadOnlyEntityClasses
         /// <summary> The AccountNumber property of the Entity DhlExpressAccount<br/><br/>
         /// </summary>
         /// <remarks>Mapped on table field: "DhlExpressAccount"."AccountNumber"<br/>
-        /// Table field type characteristics (type, precision, scale, length): Int, 10, 0, 0<br/>
+        /// Table field type characteristics (type, precision, scale, length): BigInt, 19, 0, 0<br/>
         /// Table field behavior characteristics (is nullable, is PK, is identity): false, false, false</remarks>
-        public System.Int32 AccountNumber { get; }
+        public System.Int64 AccountNumber { get; }
         /// <summary> The ShipEngineCarrierId property of the Entity DhlExpressAccount<br/><br/>
         /// </summary>
         /// <remarks>Mapped on table field: "DhlExpressAccount"."ShipEngineCarrierId"<br/>
