+++ Code/ShipWorks.Res/Data/Administration/Scripts/Update/5.19/5.19.0.0.sql
@@ -17,7 +17,7 @@ CREATE TABLE [dbo].[DhlExpressAccount]
 (
 [DhlExpressAccountID] [bigint] NOT NULL IDENTITY(1102, 1000),
 [RowVersion] [timestamp] NOT NULL,
-[AccountNumber] [int] NOT NULL,
+[AccountNumber] [bigint] NOT NULL,
 [ShipEngineCarrierId] [nvarchar] (12) COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
 [Description] [nvarchar] (50) COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
 [FirstName] [nvarchar] (30) COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
